wip: rename SessionReaderImpl to RemoteSessionReader
